pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "probelab"
version = "1.0.0-alpha.1"
description = "Browser automation health monitoring CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "JadeL" },
]
keywords = ["web-scraping", "monitoring", "health-check", "cli", "adapter", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "typer>=0.12",
    "click>=8.1",
    "httpx>=0.27",
    "selectolax>=0.3",
    "jsonschema>=4.21",
    "rich>=13.7",
    "pyyaml>=6.0",
    "tomli>=2.0;python_version<'3.11'",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.40",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.scripts]
probelab = "probelab.cli_v1:main"
probelab-legacy = "probelab.cli:main"

[project.urls]
Homepage = "https://github.com/MinuteMighty/probelab"
Repository = "https://github.com/MinuteMighty/probelab"
Issues = "https://github.com/MinuteMighty/probelab/issues"
